chore(tests): update test file paths and imports for utils reorganization
- Move encryption test files from helpers/encrypt to utils/encrypt
- Move numbers test files from helpers/numbers to utils/numbers
- Move strings test files from helpers/strings to utils/strings
- Move utilities test files from helpers/utilities to utils/utilities
- Update import paths in test files to reflect new module locations
- Update documentation references to point to utils modules
- Update script imports to use new utils paths
