+++ tests/unit/utils/encrypt/hybrid.test.js
