+++ tests/unit/utils/encrypt/utilities.test.js
