+++ tests/unit/utils/numbers/percentage.test.js
