+++ tests/unit/utils/strings/advanced-processing.test.js
@@ -2,7 +2,7 @@
 // ADVANCED STRING PROCESSING - UNIT TESTS
 // =============================================================================
 
-const stringsHelper = require('../../../../helpers/strings.helper');
+const stringsHelper = require('../../../../utils/strings.util');
 const { cerror } = require('../../../../helpers/debug.helper');
 
 // Mock the debug helper
