+++ tests/unit/utils/utilities/array.test.js
