+++ tests/unit/utils/utilities/functional.test.js
